requirements.txt
pytest
pytest-cov
coveralls
pyahocorasick
pandas
numpy
